Infer working directory from JVM -Duser.dir system property.
Previously, the working directory for forked processes was always the
base directory of the project. This caused problems for the Pants
integration since the working directory in Pants is always the workspace
directory. We can't set the project's base directory to the workspace
directory since IntelliJ requires different projects to have unique base
directories.

This commit updates the logic for inferring the working the directory of
a process by checking if the `-Duser.dir=` system property is configured
in the JVM platform configuration. If the property is customized, then
Bloop uses that working directory instead of base directory.