URL-encode file paths for manifest jars.
Previously, we used `Path.toString` for generating `Class-Path` entries
of a manifest jar. This encoding works correctly for normal paths but it
fails when the file path has special characters like `+`. To properly
handle such characters in the filepath, we need to URL-encode the file
path. This is necessary in order to use artifacts like
`bloop-backend_2.12-1.4.0-RC1+11-2b1413fb.jar`.