Support any kind of glob that is supported by `PathMatcher`.
Previously, `SourcesGlobs` post-processed the file globs from the Bloop
JSON config so that a relative glob like `*.scala` became an absolute
glob like `glob:/path/to/dir/*.scala`. Now, the file globs in the JSON
config are passed unchanged to `FileSystem.getPathMatcher` so that
it's possible to for example use the "regex:" syntax instead of "glob:"
syntax.

This change both simplifies the spec for globs while being more flexible
than the previous solution.