Transparently create a thread-safe buffer reporter for Hydra compilation
When compilation is carried out with Hydra, the same reporter instance is shared
across several threads. Hence, it is paramount for the reporter's implementation
to be thread-safe.

The existing, non thread-safe, implementation could cause the Hydra compilation
threads to crash because Reporter's problems buffer was not thread-safe (for
context, the exception was an `ArrayIndexOutOfBoundException` with index -1 -
this was caused by the concurrent calls to `Reporter.reset` made by the Scala
compiler inside `Global.compileUnits`).

This refactoring breaks both source and binary compatibility.