Hydra shouldn't load symbols from invalidated classfiles
Even though the real changes are not visible in this commit, below is an
explanation of the consequences brought by bumping the bloop-hydra-bridge.

Introduced a protected method `shouldLoadSymbolsFromBinary` in the
`HydraSymbolLoaders` to allow subclasses to decide if symbols from a binary
classfile should be loaded.

The concept of invalid classfiles only exists in Bloop, but to avoid duplicating
logic from the `HydraSymbolLoaders` we opted for introducing a method. By
default, the `HydraSymbolLoaders` always loads symbols from classfiles. But the
bloop-hydra-bridge `MergedSymbolLoaders` class overrides this default behavior
and uses the `ZincSymbolLoaders#invalidatedClassFilePaths` to decide if symbols
from a classfile should be loaded or not. Hence, the bloop-hydra-bridge version
was bumped to 0.2.0 (which requires Hydra 2.2.1).

It's worth emphasising that using Bloop with Hydra now requires Hydra 2.2.1. I
don't expect this to be an issue since there are few Hydra users at the moment
that are using Bloop.