Inline `compileSources` to avoid using `unzip`
The `unzip` call-site in Zinc's `compileSources` has
`preserveLastModified` set to `true` even if it's not needed. However,
the execution of this method produces the use of certain JNA APIs that
are not present in the last JNA version 0.5.5 used by nuprocess. As we
don't need `preserveLastModified` to be `true`, we inline the sources
and set it manually to `false`.