Reporter implementations must be thread-safe to work with Hydra
Not just the Reporter's problems buffer that must be thread-safe (Re 085159cc).

All Reporter's implementation must be thread-safe to avoid concurrency hazards
to surface when compilation is carried out with Hydra.

It's worth mentioning that if Hydra is used for compilation, `Reporter.reset` is
called by each compilation thread prior to compile the sources (see the
implementation of `Global.compileUnitsInternal`), which could cause concurrency
hazards (e.g., compilation problems already reported by some compilation thread
may be dropped). TBH, I believe this to be an extremely unlikely situation to
happen (I expect all compilation threads to have called `Reporter.reset` before
a compilation error is reported). But, since a race-condition does exist, we
will work on a fix. Note that the fix will happen on the Hydra side, since the
call to `Reporter.reset` is made by Global). Therefore, there is no action to
take in Bloop.