Remove versions in Docusaurus
The use of versions in Docusarus hasn't been really useful so far and
now that the instructions for the build tools are going to be inlined in
the actual docs it prevents us from showing always the latest version.

As a result, I disable them. I will think if we add some links to
release notes to make them easier to be discovered. Another option is to
have the blog inline the release notes.