Report test projects without test frameworks
Bloop used to report an error when running the `test` task against a
project that has an empty list of configured test frameworks. This
behavior is not convenient for targets that do not have any test
framework configured, and for which Bloop cannot infer what project (if
any) is the test variant.

This commit changes the test task, so that an error is reported if it is
run against a project that has the test or integration-test tags, and
that has no test frameworks configured. When run against a project that
is not a test project, and that has no test frameworks configured, the
test task will report a warning indicating that no test frameworks were
found.