Use coursier in all supported installation methods
This commit changes the guts of the release utils used in sbt to install
bloopgun from coursier directly. In those installation methods that
support it, we install the autocompletions as well. All the other
resources such as systemd or service files are gone because bloop now
should be managed via the bloop CLI. There's no point in managing it
with systemd or other system because bloop can be started and exited
based on the requirements of the bloop clients using the launcher.