Make `bloop about` and `bloop help` always succeed.
Previously, running `bloop help` or `bloop about` in the `$HOME`
directory could fail to run when the workspace contained invalid
`.bloop/*.json` config files.  This commit refactors the `bloop help`
and `bloop about` commands so that they print out a message and
don't attempt to load the workspace JSON files.