Register Scala generated class files before compiling Java (#1254)
Classfiles originating from Scala recompiled sources would be marked as
invalidated until after the Analysis object is created. This caused
issues with the `BloopForkedJavaCompiler`, which creates invalid
classfiles, since these invalid classfiles would be chosen over the new
ones, thus aborting compilation.

This commit changes the analysis callbacks used by Bloop so that the
generated classfiles are pushed to the classfile manager as soon as they
are known. Thanks to this change, no invalid classfiles will be created
for the recompiled sources.