Support links in path to javac (#1258)
We were using a forked javac process to compile Java sources in cases
where the `java.home` properties and the JVM platform of a project were
different, but pointed to the same file. Instead of comparing the two
paths to decide whether to fork, this commit compares the two files that
are pointed to by the paths.