Don't consider CLI classes directory an orphan directory
This commit fixes a bug in the logic deleting the orphan bsp classes
directories that was not updated when the CLI client classes dir was
moved to behave like just a normal bsp client classes dir.

This bug could have produced important problems such as
`ClassNotFoundException`s when starting a new connection while a bloop
CLI command is running.

I don't think this issue is directly related with the metals connection
issues but it might.