Catch compiler error early in the compile pipeline
When our compile implementation throws unexpectedly because of some of
the additional tasks we're running before/after hitting `CompileTask`,
we trap whatever error is thrown so that we guarantee that the
compilation will always return a value. We're assuming this assumption
holds in our logic deduplicating compilation, so this change honors it.

Fixes https://github.com/scalacenter/bloop/issues/1140