Release artifacts with Java 8, not Java 11
Fixes https://github.com/scalacenter/bloop/pull/1277

The 1.4.0 is broken for those folks using Java 8 as their default Java
configuration. This error was caused when upgrading to graalvm@20.0.0.
The previous GraalVM 19.2.0 version used **Java 8** but that changes in
the new GraalVM versions where the default in jabba pulls Java 11.

The previous attempt at fixing this release issue in
https://github.com/scalacenter/bloop/pull/1277 was a red herring.