Don't expand resource directory entries
Previously, the Bloop BSP server recursively walked all directory
resource entries and returned individual files. Now, the BSP server
returns the URIs of the directories directly and leaves it up to the
client to recursively walk the directories, if necessary. There's
nothing in the BSP spec that says that resource URIs cannot be
directories. The current behavior looses information because client
won't be able to relativize resource items by their enclosing directory.