Workaround Windows publishing issue
The native image binaries publishing on Windows is currently broken
because of an issue where sbt fails to find
org.scala-lang.modules#scala-xml_2.12;1.2.0!scala-xml_2.12.jar when sbt
is started with `cmd` rather than `bash`.

To workaround this issue, we launch sbt a first time with `bash`, and
then do the publishing with `cmd`.

Why it fails with `cmd` is a mystery to me.