Ensure we don't hash directory when using globs
Globs allow us to define the maximum depth to search for source files.
When we reach this maximum depth, the `Files.walkFileTree` will call
`visitFile` on directories that cannot be traversed.

Previously, we were adding these directories to the list of items to
hash, which would cause an exception to be thrown.