Destroy process group alongside process
Background
In Unix systems, the `kill` syscall kills only a single process. This
means, it's subprocesses are not stopped. Instead, they are
re-parented and they continue to run.

A way to manage lifetime of a set of processes is to form a Process
Group and then kill the whole group. Unfortunatelly, the
`NuProcess.start` method currently does not support Process Groups.
A way to workaround this is to delegate group creation to the process
itself and handle both cases: if it creates a group, and if it does not.

Nevertheless, we may utilize the fact that, when a process creates
a Process Group, the group's ID is equal to the PID. So the Forker may
always try to kill the group of that id. If the process created a group, then it would
be killed in total. If not, then nothing would happen. It is still
better than just killing the root process and allow children to be re-parented.

References
https://en.wikipedia.org/wiki/Orphan_process
https://man7.org/linux/man-pages/man2/kill.2.html