Ignore empty directory layout in external classes dir
Because the proces owning external classes directories might force the
creation of certain empty directories, the process responsible for
copying directories would consider the external client classes directory
was not empty and would not copy class files from the internal classes
directory.

This commit fixes that and ignores empty directories in the external
classes dir, resulting in the copying process working as expected.