Strip ANSI color escapes when colors are disabled
Previously, the `BloopLogger` would remove the colored prefix of the log
messages when colors were disabled via `--color` or `NO_COLOR`. However,
color that may be present in the log output (inserted by the running
application or a test framework for instance) would not be removed.

This commit preprocesses the log messages so that the ANSI colors are
removed from the log messages.

Fixes #1119