Fix #1358: Run test classes only once
Bloop was collecting all the test fingerprint, and creating a test task
for every matching fingerprint. This meant that test classes would be
run as many times as they have matching fingerprints.

Bloop now follows the same behavior as sbt, which will run the test with
the first matching framework.

Fixes #1358