Use our fork of test-agent
The sbt test-agent recently received a fix for duplicated tasks being
run (see https://github.com/sbt/sbt/pull/5800). We would like to start
using this fix in Bloop, but don't want to wait for the next release.

This commit adds a new artifact, called `test-agent`, which contains the
fixed version of the test-agent, so that we can use it without waiting
for sbt to publish a new version.