Fix issues with meta builds not resolving sbt artifacts
It seems that due to https://github.com/sbt/sbt/pull/6085/files we will no longer resolve sbt artifacts when running with coursier, which is the default.
sbt will instead try to use artifacts from the launcher, which does not seem to work in case of plugins.
There is an existing issue https://github.com/sbt/sbt/issues/6203, but in the meanitime we can force resolution using:

```scala
Keys.reresolveSbtArtifacts := true
```