Fix issues when starting console within workspaces that include a space
When running the `console` command an additional `--out-file` parameter is added by Bloopgun, which Bloop will write an ammonite command to. This command would hover be written as arguments separated by spaces, which makes it not work in case of workspaces containing spaces. Here, I change it to newline to avoid the issue.