Switch to the original fork of zinc and remove pipelining.
In order to make upgrading versions easier we decided to switch back to sbt/zinc instead of a custom fork. The original reason for using the fork was to enable build pipelining, but that itself has been implemented in the orginal fork.

Unfortunately, to make the migration easier we needed to remove build pipelining for now to later add it with the default mechanism.

Related to https://github.com/scalacenter/bloop/issues/1383

The benchmarks are being run to make sure we are not regressing in performance.