Fix issues with scala-collection-compat missing in bloopgun
Fixes https://github.com/scalacenter/bloop/issues/1463

scala-collection-compat is brought in by coursier and since `scala` is added to be ingored when shading, it never ends up properly on the classpath of bloopgun. This fixes the issues by adding the dependency directly. It's binary compatible, so it should not be an issue in case of conflicts.