Update sbt to 1.5.5 and remove offloading
Offloading uses internal sbt API, which means there will be issues with any update that comes up. Especially since this was never finished, I think it's better to remove it and push the current state to a separate branch.

The sbt version bump is needed to help with running the benchamarks in zinc update PR (older sbt versions keep hanging)