Fix JS compilation errors
Config module is cross published for JVM and JS, which means that config-test is also compiled with both. This is causing compilation errors for `BasicConfigSuite`, although it is never actually used in tests for js, which is why it wasn't failing on CI. This PR is just to get rid of that annoying error.