Continue if `bloopGenerate` fails for a project
I'm in a situation where I want to generate bloop files for all projects for all scala variants.

Unfortunately, if we have the following project graph

```
lazy val a = project.settings(crossScalaVersions := List("2.13.0", "3.0.0"))
lazy val b = project.settings(crossScalaVersions := List("2.13.0")).dependsOn(a)
```
and then `+bloopInstall` may fail for `b` when generating for scala `3.0.0`, because it'll then use the `3.0.0` version of `a` and `2.13.0` of `b`.

I have seen errors like this:
```
[error] Modules were resolved with conflicting cross-version suffixes in ProjectRef...:
[error]    org.scala-lang.modules:scala-collection-compat _3, _2.13
[error]    com.lihaoyi:sourcecode _2.13, _3
[error]    org.typelevel:simulacrum-scalafix-annotations _2.13, _3
[error]    org.typelevel:cats-kernel _2.13, _3
[error]    com.lihaoyi:fansi _3, _2.13
[error]    org.typelevel:cats-core _2.13, _3
```