Fix copying tasty files on incremental compilation
Previously, we removed both generatedClassFile and rebasedClassFile from invalidations, but we only did it for rebased product files such as `.tasty`. Now we do it for both

I am not 100% sure why sometimes invalidations happen for rebased and sometimes for generated files, but that can happen according to the comment above.

What happened here is that the generated files existed, but the rebased one was in the invalidation list, so later on we would skip it.