Add support for source generators in Bloop
This commit changes the Bloop project definition to add a new field,
`sourceGenerators`, which can be used to define source generators.

Source generators are parts of the build that are run when Bloop gathers
the sources of a project (either when queried via BSP or when
compiling).

Source generators are configured with globs matching input files, an
output directory, and a program to run. When the inputs or known outputs
of a source generator change, the source generator is automatically
re-run. When there are no changes, the result of a source generator is
cached.

The command to execute to run a source generator will be built from the
configured `argv`, followed by the output directory and the input files.