chore: Remove java-debug and ignore updates for nailgun and release-early
javaDebug is brought in by scala-debug-adapter, so no need to define it here.

Nailgun version should fixed, since it's the one Bloop works with (dropping the fork would be awesome.)

Release-early should be replaced by sbt-ci-release so don't want to change the version and risk regression. Especially since it's not actively maintained.