refactor: rework/rethink the meta meta build stuff
Alright, so I'm pretty confident with this at the moment, and quite
happy with it. Much of it was mimicking what Alex did in
bloop-core by just ripping it all out.

As I stated in the initial description, it seems that the entire notion
of having the sbt-bloop-build-naked and the bloop-shaded-plugin is
primarily to dogfood itself in a way that doesn't conflict with the
issue that is outlined in sbt/sbt#892. Maybe at one point in time when
we were rapidly developing the sbt plugin and build and always wanting
to dogfood those changes this may have been worth it, but I don't think
it is anymore. The entire meta-meta build can just be thrown away if we
get rid of the idea of wanting to do this. Instead, we just ensure that
we publish sbt-bloop before situations where we want to ensure we're
using that. This is primarily when we're running the frontend tests and
we an the Bloop definitions of the projects in the frontend resources
and when we're running buildpress. In both of these scenarios we just
publish sbt-bloop first, and it works the same way. Benchmarks will also
work exactly the same as the version is passed directly to it that we're
using. We included the sbt-bloop-build-naked there before, but from my
understanding it was again only to dogfood the bloop generation when we
can just use the one we just published.

So to outline the changes, this

- gets rid of the sbt-bloop-build-naked
- gets rid of the bloop-build-shaded-plugin
- benchpress is still able to run fine
- frontend test resources are still able to be generated fine
- tests all still pass
- the build is way less crazy

The main thing we lose out on is:

We no longer dogfood the absolute latest changes This is a tradeoff I'm
100% ok with.