refactor: stop shading launcher and bloopgun
Looking back I can't find a reason as to why this was done, so my
assumption is that this was done as a precaution. Looking through the
deps they are quite minimal. Testing around with Metals using the
unshaded version of launcher doesn't cause any issues. I also see that
multiple apps are using bloop-rifle instead, and that's also not shaded,
arguably has more deps, and isn't causing issues. If this is ok, we can
get rid of the entire custom shading logic we have that really
simplifies things and also allows us to no longer publish both shaded
and not shaded versions of launcher and bloopgun. If we detect problems,
we can easily then bring in sbt-shading.