refactor(ci): migrate to sbt-ci-release
This pr changes the way we do releases and migrates us away from the
forked sbt-release-early that's no longer maintained and is also
bringing in some old versions of scala-xml which will be problematic
when we try to update our sbt stuff.

Plus this simplifies the release a little bit. The one difference is
that we'll now be publishing actual `-SNAPSHOT` into sonatype snapshots.