bugfix: Properly clean up orphan directories for internal classes
It seems that the genericClassesDir is not being used at all by the compiler, so cleaning orphan directories in it was not achieving anything.

We might need to figure out how to use `genericClassesDir` and whether it needs to be used at all, but for now this fixes the issue of large .bloop directory.