Support Scala 3's Best Effort compilation
This commit allows bloop to support best effort compilation for Metals,
with it being enabled with the other Metals options.

Best Effort is meant to be a set of Scala 3 compiler options that allow
the errored, but typed programs, to be able to be serialized into a
TASTy aligned format (Best Effort TASTy), and later to reuse those
typed program trees in both the dependent projects, and in the
presentation compiler.

Those best effort tasty files are serialized to a
META-INF/best-effort in classesDir. Best effort compilation
may fail, similarly to the regular compilation. In that case we
stop compiling the downstream dependencies.

Currently best-effort compilation is unable t assist in producing the
zinc analysis files, which lead to the projects being recompiled every
time. This is solved with a custom hashing solution.