Support Scala 3's Best Effort compilation
This commit allows bloop to support best effort compilation for Metals,
with it being enabled with the other Metals options.

Best Effort is meant to be a set of Scala 3 compiler options that allow
the errored, but typed programs, to be able to be serialized into a
TASTy aligned format (Best Effort TASTy), and later to reuse those
typed program trees in both the dependent projects, and in the
presentation compiler.

Those best effort tasty files are always serialized to a seperate
directory, which in this PR is managed by bloop. Best effort compilation
may fail, similarly to the regular compilation.

The best effort directory is kept outside of the regular build
directory, in `.bloop/project-id/scala-3/best-effort`. There, two
directories are managed, `build` and `dep`. Whenever metals compiles
a project the `build` directory is populated with best effort artifacts.
Then, if the best effort compilation is successful, they are copied to
the 'dep' directory, which is used for dependencies, metals presentation
compiler etc. This way if a compilation fails, the artifacts from the
previous compilation can still be kept. This is not completely
dissimilar to how the regular compilation works here, with the
difference being the more straigthforward implementation, without any
counters that monitor how many tasks are being executed with the `dep`
directory concurrently.