chore: Adjust Mill build file to fix Bloop compilation
Previously, if importing build as Bloop we would not get proper classpath for bridge modules as they would be missing tansitive classes directories. This is probably a bug in mill bloop plugin, but I am not sure we want to dig in there currently. Now, I added explicit depndency on those moduls which seems to have fixed compilation.