refactor: don't send task start/end for no-op compilations
This might be a bit opinionated and I'm not 100% sure on the reason this
was done in the first place, but this changes the current behavior
around reporting start and end compilations for no-ops. This still
retains the compile report if it was a no-op and also still checks the
diagnostics and reports them, but skips the task start/end.

Originally this was added in a commit without any context... but there
was a comment that said:

> // When no-op, we keep reporting the start and the end of compilation for consistency

However, I'm not really sure that consistency matters here. In reality
this ends up creating a bunch of noise on the client side, especially
when these tasks turn into LSP progress notifications that aren't useful
for the user to see. You can see more context about this change in the
issue reported [here](https://github.com/scalameta/metals/issues/6099)
and also the discussion found
[here](https://github.com/build-server-protocol/build-server-protocol/discussions/654).
It also seems that in some projects like scala-cli these are even [ignored](https://github.com/VirtusLab/scala-cli/blob/6b7a10007e4eefde717079255e0df38c027f788b/modules/build/src/main/scala/scala/build/ConsoleBloopBuildClient.scala#L109).