bugfix: Fix CI by updating JMH and inlining sbt-github-release
Both of the dependencies don't seem to be available on maven central which is why the build is currently failing

For JMH I fixed the compilation by inlining some missing classes and fixing enum. I am not 100% sure if that is correct, but seems to work ok.

For github release I just inlined the one file.