feature: Switch to using Bloop Rifle and backport all improvements
Scala CLI uses a fork of Bloop by default and that fork has added a bunch of improvements especially connected to using domain sockets, which should work much nicer in our case. IT also solves the issue of having multiple Bloop servers on one machine, but with different users.

This might sometimes lead to using 2 bloop servers on one machine, which take up memory uneccesarily. To fix that I decided to backport most of the fixes from the fork, add bloop rifle and deprecate launcher.