bugfix: Try to stop bloop if unable to run about
This was raised on discord and it seems when we fail to parse the output of `about` everything will break.

This is because there is a server running and we will not restart it and the output of about command will again we wrong when we reask.

Now, we should try to exit it first in that case since msot likely the server is broken.