Add --link flag to compile to allow linking through BSP
Before this commit linking could only be done through the run command with BSP.
Because the BSP protocol does describe a linking command (as it is often a part
of the compile step) the flag --link has been added to the compile command. This
is useful for those kinds of projects that need the linker to be run to produce
source code for further toolchains, e.g. a ScalaJS project might often use
@JSExport and then run the code from some other JS entrypoint.