Set up test for bsp linking
Removes the dependency on frontend in jsbridge to avoid circular dependency when
publishing bridge before running tests in frontend.

The only reason jsbridge dependend on frontend was for logging the project name
in jsbridge1, for jsbridge06 only the project name was needed for setting the
output dir. frontend was also needed for running the tests for the bridge.
Moving the bride test to it's own module let's us keep the dependecy graph neat
and tidy.