Better error reporting for the linking stage and NoOp for linking JVM
When linking a JVM project the link task now simply noops. This lets
clients send all build projects to the link task and the task will only
link where possible. This way the client does not need to be explicit
when linking (although still possible).

Errors when linking are now accumulated and error logged. Bsp returns an
error code when encountering linking errors.