Added the flags --debug and --release to bspCompile for linking
This lets the client override the mode without making changes to the
model. This is more in line with how fullLink/fastLink works in sbt.

Using flags through BSP for this lets the client override the mode
without having to change the actual build. This is very useful for this
task.