Fix issue with frequent file saves leading to leftover errors
This was caused by the custom hashing mechanism used for best effort
compilations where we would needlessly hash the inputs twice - once
before compilation (to see it is necessary), and once after
(to save the results). If a file was resaved during the compilation,
by the end of it the hash input would be outdated and unreleted to the
hash output, so bloop would cease recompilations until next change in
the source file, leading to incoherent errors.

So now, we decouple the output hashing from the input hashing, and only
hash the outputs after the compilation, and inputs always before the
compilation. In theory, this could cause an unnecessary recompilation
if a file is re-saved after reading inputs, but before compiling leading
to a needless compilation, but this is the case where too much is better
then not enough.