Move betasty directory cleaning into the main thread
Previously it was done in backgroundTasks, which could cause issues
when combined with the hashing mechanism, itself in the main thread.
This race condition could cause the iterator on the files would throw an
error, and downstream compialations would switch to the best effort
mode, even if the compilation should be successful.